requirements.txt
streamlit
pandas
openpyxl
reportlab
arabic-reshaper
python-bidi